+++ data_validation/data_validation.py
@@ -13,12 +13,12 @@
 # limitations under the License.
 
 import json
-import logging
 import warnings
 
 import ibis.backends.pandas
 import numpy
 import pandas
+import logging
 
 from data_validation import combiner, consts, metadata
 from data_validation.config_manager import ConfigManager
@@ -107,6 +107,7 @@ class DataValidation(object):
 
         # Filter for only first primary key (multi-pk filter not supported)
         primary_key_info = self.config_manager.primary_keys[0]
+
         query = RandomRowBuilder(
             [primary_key_info[consts.CONFIG_SOURCE_COLUMN]],
             self.config_manager.random_row_batch_size(),
@@ -114,9 +115,12 @@ class DataValidation(object):
             self.config_manager.source_client,
             self.config_manager.source_schema,
             self.config_manager.source_table,
+            self.validation_builder.source_builder,
         )
 
         random_rows = self.config_manager.source_client.execute(query)
+        if len(random_rows) == 0:
+            return
         filter_field = {
             consts.CONFIG_TYPE: consts.FILTER_TYPE_ISIN,
             consts.CONFIG_FILTER_SOURCE_COLUMN: primary_key_info[
@@ -274,8 +278,8 @@ class DataValidation(object):
             schema_index.append(key)
         pd_schema = pandas.Series(schema_data, index=schema_index)
         if verbose:
-            print("-- ** Pandas Schema ** --")
-            print(pd_schema)
+            logging.info("-- ** Pandas Schema ** --")
+            logging.info(pd_schema)
 
         return pd_schema
 
@@ -355,12 +359,12 @@ class DataValidation(object):
                 )
             except Exception as e:
                 if self.verbose:
-                    print("-- ** Logging Source DF ** --")
-                    print(source_df.dtypes)
-                    print(source_df)
-                    print("-- ** Logging Target DF ** --")
-                    print(target_df.dtypes)
-                    print(target_df)
+                    logging.error("-- ** Logging Source DF ** --")
+                    logging.error(source_df.dtypes)
+                    logging.error(source_df)
+                    logging.error("-- ** Logging Target DF ** --")
+                    logging.error(target_df.dtypes)
+                    logging.error(target_df)
                 raise e
         else:
             result_df = combiner.generate_report(