+++ data_validation/query_builder/random_row_builder.py
@@ -14,24 +14,27 @@
 
 import random
 import logging
+from typing import List
+from io import StringIO
+import sqlalchemy as sa
 import ibis
 import ibis.expr.operations as ops
 import ibis.expr.types as tz
 import ibis.expr.rules as rlz
 import ibis.backends.base_sqlalchemy.compiler as sql_compiler
+import ibis.backends.base_sqlalchemy.alchemy as sqla
+import ibis.backends.pandas.execution.util as pandas_util
+
 from ibis_bigquery import BigQueryClient
 from ibis.backends.impala.client import ImpalaClient
 from ibis.backends.pandas.client import PandasClient
-import ibis.backends.pandas.execution.util as pandas_util
-
+from ibis.backends.postgres.client import PostgreSQLClient
 from ibis.expr.signature import Argument as Arg
-from typing import List
 from data_validation import clients
-from io import StringIO
+from data_validation.query_builder.query_builder import QueryBuilder
 
 
 """ The QueryBuilder for retreiving random row values to filter against."""
-
 
 ######################################
 ### Adding new datasources should be
@@ -44,7 +47,10 @@ from io import StringIO
 RANDOM_SORT_SUPPORTS = {
     PandasClient: "NA",
     BigQueryClient: "RAND()",
+    clients.TeradataClient: None,
     ImpalaClient: "RAND()",
+    clients.OracleClient: "DBMS_RANDOM.VALUE",
+    PostgreSQLClient: "RANDOM()",
 }
 
 
@@ -85,7 +91,11 @@ class RandomRowBuilder(object):
         self.batch_size = batch_size
 
     def compile(
-        self, data_client: ibis.client, schema_name: str, table_name: str
+        self,
+        data_client: ibis.client,
+        schema_name: str,
+        table_name: str,
+        query_builder: QueryBuilder,
     ) -> ibis.Expr:
         """Return an Ibis query object
 
@@ -95,7 +105,9 @@ class RandomRowBuilder(object):
             table_name (String): The name of the table to query.
         """
         table = clients.get_ibis_table(data_client, schema_name, table_name)
-        randomly_sorted_table = self.maybe_add_random_sort(data_client, table)
+        compiled_filters = query_builder.compile_filter_fields(table)
+        filtered_table = table.filter(compiled_filters) if compiled_filters else table
+        randomly_sorted_table = self.maybe_add_random_sort(data_client, filtered_table)
         query = randomly_sorted_table.limit(self.batch_size)[self.primary_keys]
 
         return query
@@ -105,6 +117,10 @@ class RandomRowBuilder(object):
     ) -> ibis.Expr:
         """Return a randomly sorted query if it is supported for the client."""
         if type(data_client) in RANDOM_SORT_SUPPORTS:
+            # Teradata 'SAMPLE' is random by nature and does not require a sort by
+            if type(data_client) == clients.TeradataClient:
+                return table
+
             return table.sort_by(
                 RandomSortKey(RANDOM_SORT_SUPPORTS[type(data_client)]).to_expr()
             )
@@ -191,3 +207,41 @@ def format_order_by(self):
 
 
 sql_compiler.Select.format_order_by = format_order_by
+
+#######################################
+##### Override Order By for SQL Alchemy
+#######################################
+
+
+def _add_order_by(self, fragment):
+    if not len(self.order_by):
+        return fragment
+
+    clauses = []
+    for expr in self.order_by:
+        key = expr.op()
+        sort_expr = key.expr
+
+        ##### ADDING CODE TO FORMAT #####
+        if isinstance(expr, RandomSortExpr):
+            arg = sa.sql.literal_column(sort_expr.op().value)
+            clauses.append(arg)
+            continue
+        ##### END ADDING CODE TO FORMAT #####
+
+        # here we have to determine if key.expr is in the select set (as it
+        # will be in the case of order_by fused with an aggregation
+        if sqla._can_lower_sort_column(self.table_set, sort_expr):
+            arg = sort_expr.get_name()
+        else:
+            arg = self._translate(sort_expr)
+
+        if not key.ascending:
+            arg = sa.desc(arg)
+
+        clauses.append(arg)
+
+    return fragment.order_by(*clauses)
+
+
+sqla.AlchemySelect._add_order_by = _add_order_by