+++ data_validation/schema_validation.py
@@ -62,7 +62,6 @@ class SchemaValidation(object):
                 "source_agg_value",
                 "target_agg_value",
                 "validation_status",
-                "error_result.details",
             ],
         )
 
@@ -93,9 +92,6 @@ class SchemaValidation(object):
         )
         df.insert(loc=10, column="aggregation_type", value="Schema")
 
-        # TODO: not being displayed/stored at the moment, but it would be useful
-        del df["error_result.details"]
-
         # empty columns added due to changes on the results schema
         df.insert(loc=14, column="primary_keys", value=None)
         df.insert(loc=15, column="num_random_rows", value=None)
@@ -121,25 +117,23 @@ def schema_validation_matching(source_fields, target_fields, exclusion_fields):
 
     if exclusion_fields is not None:
         for field in exclusion_fields:
-            del source_fields_casefold[field]
-            del target_fields_casefold[field]
+            source_fields_casefold.pop(field, None)
+            target_fields_casefold.pop(field, None)
 
     # Go through each source and check if target exists and matches
     for source_field_name, source_field_type in source_fields_casefold.items():
         # target field exists
         if source_field_name in target_fields_casefold:
             # target data type matches
-            if source_field_type == target_fields_casefold[source_field_name]:
+            target_field_type = target_fields_casefold[source_field_name]
+            if source_field_type == target_field_type:
                 results.append(
                     [
                         source_field_name,
                         source_field_name,
-                        "1",
-                        "1",
+                        str(source_field_type),
+                        str(target_field_type),
                         consts.VALIDATION_STATUS_SUCCESS,
-                        "Source_type:{} Target_type:{}".format(
-                            source_field_type, target_fields_casefold[source_field_name]
-                        ),
                     ]
                 )
             # target data type mismatch
@@ -148,12 +142,9 @@ def schema_validation_matching(source_fields, target_fields, exclusion_fields):
                     [
                         source_field_name,
                         source_field_name,
-                        "1",
-                        "1",
+                        str(source_field_type),
+                        str(target_field_type),
                         consts.VALIDATION_STATUS_FAIL,
-                        "Data type mismatch between source and target. Source_type:{} Target_type:{}".format(
-                            source_field_type, target_fields_casefold[source_field_name]
-                        ),
                     ]
                 )
         # target field doesn't exist
@@ -162,10 +153,9 @@ def schema_validation_matching(source_fields, target_fields, exclusion_fields):
                 [
                     source_field_name,
                     "N/A",
-                    "1",
-                    "0",
+                    str(source_field_type),
+                    "N/A",
                     consts.VALIDATION_STATUS_FAIL,
-                    "Target doesn't have a matching field name",
                 ]
             )
 
@@ -176,10 +166,9 @@ def schema_validation_matching(source_fields, target_fields, exclusion_fields):
                 [
                     "N/A",
                     target_field_name,
-                    "0",
-                    "1",
+                    "N/A",
+                    str(target_field_type),
                     consts.VALIDATION_STATUS_FAIL,
-                    "Source doesn't have a matching field name",
                 ]
             )
     return results